requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
mysqlclient==2.2.4